# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest-Api for library",
        "title": "BibliotecaAPI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/api/admin/loans": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos los préstamos activos en un rango de fechas específico para administradores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Obtener préstamos activos por rango de fechas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha de inicio (YYYY-MM-DD)",
                        "name": "startdate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha de fin (YYYY-MM-DD)",
                        "name": "enddate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Lista de préstamos activos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parámetros de fecha faltantes o inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "No tienes permisos para acceder a este recurso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/loans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos los préstamos activos asociados a un usuario por su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Obtener préstamos activos de un usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "usuario_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Lista de préstamos activos del usuario",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parámetro 'usuario_id' faltante o inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No hay préstamos activos para este usuario",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo préstamo para un usuario autenticado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Crear un nuevo préstamo",
                "parameters": [
                    {
                        "description": "Crear prestamo",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Loan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Préstamo creado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud mal formada o parámetros faltantes",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "No autorizado: token JWT inválido o ausente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/loans/completed": {
            "get": {
                "description": "Get completed loan history for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Get user completed loan history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "idsocio",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Loan"
                            }
                        }
                    }
                }
            }
        },
        "/api/loans/extend/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extend the return date of a loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Extend loan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New return date",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loan extended successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/loans/history": {
            "get": {
                "description": "Get loan history for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Get user loan history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "idsocio",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Loan"
                            }
                        }
                    }
                }
            }
        },
        "/api/loans/return/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a loan as returned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestamos"
                ],
                "summary": "Return loan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loan returned successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Loan": {
            "type": "object",
            "properties": {
                "fecha_devolucion": {
                    "type": "string"
                },
                "fecha_prestamo": {
                    "type": "string"
                },
                "libro_id": {
                    "type": "integer"
                },
                "usuario_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}